Revert "Enable Student Work detail views"
