+++ source/navigation.js
@@ -14,7 +14,7 @@ import {FilterView} from './views/components/filter'
 import NewsView from './views/news'
 import NewsItemView from './views/news/news-item'
 import SISView from './views/sis'
-import {StudentWorkDetailView} from './views/sis/student-work-carls'
+// import JobDetailView from './views/sis/student-work/detail'
 import {
   BuildingHoursView,
   BuildingHoursDetailView,
@@ -68,6 +68,7 @@ export const AppNavigator = StackNavigator(
     FaqView: {screen: FaqView},
     FilterView: {screen: FilterView},
     HelpView: {screen: HelpView},
+    // JobDetailView: {screen: JobDetailView},
     LegalView: {screen: LegalView},
     MenusView: {screen: MenusView},
     NewsItemView: {screen: NewsItemView},
@@ -80,7 +81,6 @@ export const AppNavigator = StackNavigator(
     StreamingView: {screen: StreamingView},
     StudentOrgsDetailView: {screen: StudentOrgsDetailView},
     StudentOrgsView: {screen: StudentOrgsView},
-    StudentWorkDetailView: {screen: StudentWorkDetailView},
     TransportationView: {screen: TransportationView},
     BusMapView: {screen: BusMapView},
     XyzBusView: {screen: ConnectedXyzBusView},
