+++ source/views/sis/index.js
@@ -7,7 +7,7 @@
 import {TabNavigator} from '../components/tabbed-view'
 
 import BalancesView from './balances'
-import {StudentWorkView} from './student-work-carls'
+import StudentWorkView from './student-work'
 // import CoursesView from './courses'
 // import SearchView from './search'
