--- source/views/sis/student-work-carls/index.js
@@ -1,4 +1,0 @@
-// @flow
-
-export {StudentWorkView} from './list'
-export {StudentWorkDetailView} from './detail-wrapper'
