+++ source/views/sis/student-work/index.js
@@ -6,25 +6,20 @@
  */
 
 import React from 'react'
-import {StyleSheet, Text, SectionList} from 'react-native'
+import {StyleSheet, Text, FlatList} from 'react-native'
 import {TabBarIcon} from '../../components/tabbar-icon'
 import type {TopLevelViewPropsType} from '../../types'
 import * as c from '../../components/colors'
-import {ListSeparator, ListSectionHeader} from '../../components/list'
+import {ListSeparator} from '../../components/list'
 import {tracker} from '../../../analytics'
 import bugsnag from '../../../bugsnag'
 import {NoticeView} from '../../components/notice'
 import LoadingView from '../../components/loading'
 import delay from 'delay'
-import toPairs from 'lodash/toPairs'
-import orderBy from 'lodash/orderBy'
-import groupBy from 'lodash/groupBy'
-import {toLaxTitleCase as titleCase} from 'titlecase'
 import {JobRow} from './job-row'
-import type {JobType} from './types'
+import type {ThinJobType} from './types'
 
-const jobsUrl =
-  'https://www.stolaf.edu/apps/stuwork/index.cfm?fuseaction=getall&nostructure=1'
+const jobsUrl = 'https://apps.carleton.edu/campus/sfs/employment/feeds/jobs'
 
 const styles = StyleSheet.create({
   listContainer: {
@@ -32,17 +27,20 @@ const styles = StyleSheet.create({
   },
 })
 
+const fetchJobs = (): Array<ThinJobType> =>
+  fetchXml(jobsUrl).then(resp => resp.rss.channel.item)
+
 export default class StudentWorkView extends React.PureComponent {
   static navigationOptions = {
-    headerBackTitle: 'Open Jobs',
-    tabBarLabel: 'Open Jobs',
+    headerBackTitle: 'Job Postings',
+    tabBarLabel: 'Job Postings',
     tabBarIcon: TabBarIcon('briefcase'),
   }
 
   props: TopLevelViewPropsType
 
   state: {
-    jobs: Array<{title: string, data: Array<JobType>}>,
+    jobs: Array<ThinJobType>,
     loaded: boolean,
     refreshing: boolean,
     error: boolean,
@@ -59,28 +57,8 @@ export default class StudentWorkView extends React.PureComponent {
 
   fetchData = async () => {
     try {
-      const data: Array<JobType> = await fetchJson(jobsUrl)
-
-      // force title-case on the job types, to prevent not-actually-duplicate headings
-      const processed = data.map(job => ({...job, type: titleCase(job.type)}))
-
-      // Turns out that, for our data, we really just want to sort the categories
-      // _backwards_ - that is, On-Campus Work Study should come before
-      // Off-Campus Work Study, and the Work Studies should come before the
-      // Summer Employments
-      const sorted = orderBy(
-        processed,
-        [
-          j => j.type, // sort any groups with the same sort index alphabetically
-          j => j.office, // sort all jobs with the same office
-          j => j.lastModified, // sort all jobs by date-last-modified
-        ],
-        ['desc', 'asc'],
-      )
-
-      const grouped = groupBy(sorted, j => j.type)
-      const mapped = toPairs(grouped).map(([title, data]) => ({title, data}))
-      this.setState(() => ({jobs: mapped}))
+      const jobs = await fetchJobs()
+      this.setState(() => ({jobs}))
     } catch (err) {
       tracker.trackException(err.message)
       bugsnag.notify(err)
@@ -105,16 +83,13 @@ export default class StudentWorkView extends React.PureComponent {
     this.setState(() => ({refreshing: false}))
   }
 
-  onPressJob = (job: JobType) => {
+  onPressJob = (job: ThinJobType) => {
     this.props.navigation.navigate('JobDetailView', {job})
   }
 
-  keyExtractor = (item: JobType, index: number) => index.toString()
+  keyExtractor = (item: ThinJobType, index: number) => index.toString()
 
-  renderSectionHeader = ({section: {title}}: any) =>
-    <ListSectionHeader title={title} />
-
-  renderItem = ({item}: {item: JobType}) =>
+  renderItem = ({item}: {item: ThinJobType}) =>
     <JobRow job={item} onPress={this.onPressJob} />
 
   render() {
@@ -127,15 +102,12 @@ export default class StudentWorkView extends React.PureComponent {
     }
 
     return (
-      <SectionList
-        ListEmptyComponent={
-          <NoticeView text="There are no open job postings." />
-        }
+      <FlatList
+        ItemSeparatorComponent={ListSeparator}
+        ListEmptyComponent={<NoticeView text="There are no job postings." />}
         keyExtractor={this.keyExtractor}
         style={styles.listContainer}
-        sections={(this.state.jobs: any)}
-        renderSectionHeader={this.renderSectionHeader}
-        ItemSeparatorComponent={ListSeparator}
+        data={this.state.jobs}
         refreshing={this.state.refreshing}
         onRefresh={this.refresh}
         renderItem={this.renderItem}
