+++ source/views/sis/student-work/types.js
@@ -1,18 +1,18 @@
 // @flow
 
-export type JobType = {
-  id: number,
-  comments: string,
-  contactEmail: string,
-  contactFirstName: string,
-  contactLastName: string,
-  contactPhone: number,
-  description: string,
-  hoursPerWeek: string,
-  lastModified: string,
-  office: string,
-  skills: string,
-  timeOfHours: string | number,
+import type moment from 'moment'
+
+export type FullJobType = {
   title: string,
-  type: string,
+  department?: string,
+  dateOpen: moment,
+  duringTerm: boolean,
+  duringBreak: boolean,
+  description: string,
+}
+
+export type ThinJobType = {
+  title: string,
+  pubDate: string, // sadly, this is just the date published
+  link: string,
 }
